[refactor] domain-search-es 모듈 제거
