+++ mono/src/main/java/woowa/team4/bff/search/controller/SearchApiController.java
@@ -19,14 +19,14 @@ import woowa.team4.bff.search.command.SearchRestaurantCommand;
 public class SearchApiController {
     private final SearchService searchService;
 
-    @GetMapping("/es")
-    public ApiResult<List<RestaurantSearchResult>> searchRestaurants(
-            @RequestParam("keyword") String keyword,
-            @RequestParam("deliveryLocation") String deliveryLocation) {
-        List<RestaurantSearchResult> response = searchService.searchEs(
-                SearchRestaurantCommand.of(keyword, deliveryLocation));
-        return success(response);
-    }
+//    @GetMapping("/es")
+//    public ApiResult<List<RestaurantSearchResult>> searchRestaurants(
+//            @RequestParam("keyword") String keyword,
+//            @RequestParam("deliveryLocation") String deliveryLocation) {
+//        List<RestaurantSearchResult> response = searchService.searchEs(
+//                SearchRestaurantCommand.of(keyword, deliveryLocation));
+//        return success(response);
+//    }
 
     @GetMapping("/like-and-join")
     public ApiResult<List<RestaurantSearchResult>> searchRestaurantsUsingLikeAndJoin(
